Fix GH-17717: Socket maximum timeout of 2147 seconds?
The fix for GH-16809[1] used a way too small timeout maximum on
Windows.  We correct this.

However, later on the timeout is applied to the current timestamp, so
we would need to take that into account, but due to the elapsed time
between the check and the actual network request, this could not be
precise, and the resulting error message would be confusing, since
after the developer would adjust the timeout to the reported maximum,
the check would fail again, now reporting a lower maximum timeout.

Thus we silently cap the value in `php_network_set_limit_time()` to
avoid undefined behavior, and are not picky about the usec value (we
just assume a second more), since there is a bigger issue, namely that
we hit the Y2038 problem on Windows.

[1] <https://github.com/php/php-src/pull/16810>