Drop special casing for IIS to cover TLS protocol error
This had been introduced to fix bug 23220[1], although it is not even
clear that IIS caused the problem (some comments mention that this
happened with Apache).

Anyhow, in hindsight it seems to be a bad idea to brush warnings about
improper implementation of the *TLS* protocol under the carpet; if a
server does not properly implement TLS, users should be aware of that,
and report that to the respective developers.

[1] <https://bugs.php.net/bug.php?id=23220>