GitHub actions: add new automation to limit directory sizes
After discovering in August that Zend/tests had over 2500 entries, and that
GitHub limits the display to only the first 1000 entries, I filed GH-15631 and
sent a whole slew of patches to split up that directory and organize the tests
a bit better. After the most recent patch, GH-17920, there are now once again
fewer than 1,000 entries in that directory (for what seems to be the first time
since the PHP 5.3 development cycle).

Add a new GitHub action to the test suite that will ensure that no directory
gets above 1,000 items - while the Zend/tests directory was the only one that
had already hit that limit, there isn't much harm in checking all of the
directories, and some of the extensions have hundreds of tests already and
could conceivably hit the limit at some point.

Closes GH-15631