Optimization for htmlspecialchars function.
A dedicated php_htmlspecialchars function instead of the “universal”
php_escape_html_entities_ex.
We work with ASCII-compatible encodings, we can employ byte-by-byte scanning and a lookup
table to identify special characters. For c < 0x80, the lookup table is used; for
potentially multi-byte characters, we continue to rely on get_next_char.
This approach provides a noticeable performance improvement for ASCII strings and some
 improvement for multi-byte strings due to more optimized logic.