Resizing a test entity.
The new htmlspecialchars function respects the maximum entity size, defined as LONGEST_ENTITY_LENGTH.
There is no strict limit on the length of a numeric entity in the HTML and XML specifications,
but in practice the maximum possible is &#x10FFFF;, which takes up 10 characters.
Any numeric entities larger than this size are effectively invalid and will not be processed by browsers.