[RFC?] Add curl websocket bindings
This adds bindings to Curl's websocket support that was added in
Curl 7.86.0. Instead of just pass-through C bindings, the API added
here is a little more high-level from a type-safety perspective.
I.e. this introduces an enum CurlWsMessageType and a DTO class
CurlWsFrame for nicer and safer API handling.

Still WIP.