Fix flaky timezone test fixture
As documented in [RunningTests], tests have to be written to be
independent of any php.ini file.

Without specifying the timezone, the var_dump() will make use of the
default timezone that may not be the expected "UTC" timezone.

This renders the test-cases flaky and will make them fail if the
`date.timezone` is different from the string "UTC".

Fix is to interpolate the timezone in the test fixture or to set
`date.timezone` to `UTC`.

[RunningTests]: docs/source/miscellaneous/running-tests.rst