Fix flaky extra mail header (GH-8086)
As documented in [RunningTests], tests have to be written to be
independent of any php.ini file.

As the tests are based on counting the line terminators, the
mail.add_x_header php.ini setting has to be off.

Fix is to set `mail.add_x_header` to `off`.

[RunningTests]: docs/source/miscellaneous/running-tests.rst
fix-up-of: GH-8086