Add DNS testing setup using BIND 9
This creates a test frameword for testing PHP DNS functions that are
currently completely untested. It provides script for starting BIND 9
DNS server that uses locally defined zone files.  Those can be then used
for creating specific test scenarios without a need to use online tests.
As part of that, the systemd resolv script is provided to use BIND as a
primary resolver. This all is then used in the CI.