Adjust static inline functions which don't inline
These are found by adding `-Winline` on gcc. The flag is accepted on
clang for compatibility with gcc, but it doesn't do anything.

Mostly, this removes `inline` from the declarations. For local static
functions, there is no point in adding `inline`. This stems from
either a misunderstanding of what it means, or alternatively is a
holdover from ancient times where it may have meant something. A
compiler is free to inline a static function, and specifying inline
doesn't make the compiler inline the function. It does however make
it harder to due analysis of failed inlining, so I remove them.

However, for select functions, this changes them `zend_always_inline`
instead. This is because these functions didn't inline, but probably
should for performance reasons. For instance, `_class_exists_impl` is
used in multiple frameless function handlers. I think in these cases,
the authors were _hoping_ they'd be inlined due to fact that are
important for performance. There's also `_php_search_array`, which is
used with constant args on a variety of functions. It fails to inline
due its size, but if it _did_ inline, it would cut large amounts of
code out, so forcing the inlining seemed worth it.