Allow building opcache statically
Opcache is both a Zend extension and a PHP module. When Opcache is built
statically, the engine will attempt to initialize it as a PHP module, but will
not initialize the Zend extension.

Here I make Opcache a two-way hybrid extension [1] so that initializing it as a
PHP module also initializes the Zend extension.

[1] https://www.phpinternalsbook.com/php7/extensions_design/zend_extensions.html#hybrid-extensions